migrate accelerometer reset code to a pgReset callback method. relocate the reset code to reduce dependencies.
